Use http://localhost:8080/transaction for testing that its working.

+++ src/main/java/za/ac/cput/wisebank/repository/TransactionRepository.java
@@ -10,16 +10,16 @@ import java.util.Optional;
 
 
 @Repository
-public interface TransactionRepository extends JpaRepository<Transaction, Long> {
+public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
 
     @Override
     Transaction save(Transaction transaction);
 
     @Override
-    void deleteById(Long TransactionId);
+    void deleteById(Integer TransactionId);
 
     @Override
-    Optional<Transaction> findById(Long TransactionId);
+    Optional<Transaction> findById(Integer TransactionId);
 
     @Override
     List<Transaction> findAll();
